% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_rf.R
\name{analyze_rf}
\alias{analyze_rf}
\title{Analyze Random Forest}
\usage{
analyze_rf(model)
}
\arguments{
\item{model}{A random forest model object.}
}
\description{
Take a random forest model and extract accuracy (if classification) or MSE (if regression).
}
\examples{

library(ranger)

iris_rf <- ranger(Species ~ . , data = iris)
analyze_rf(iris_rf)
}
\keyword{analyze}
