% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_dollar_sign.R
\name{append_dollar_sign}
\alias{append_dollar_sign}
\title{Append Dollar Sign}
\usage{
append_dollar_sign(vec, regex)
}
\arguments{
\item{vec}{A vector}

\item{regex}{A regex of elements in that vector to append the dollar sign to.}
}
\description{
Tack on a dollar sign to the end of values in certain columns.
}
\examples{
append_dollar_sign(names(mtcars), "hp|dr|car")
}
\keyword{dollar}
