% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beep_on_error.R
\name{beep_on_error}
\alias{beep_on_error}
\title{Trim content}
\usage{
beep_on_error(inp, error_beep = 9)
}
\arguments{
\item{inp}{Expression to be evaluated. Cannot be piped (yet)}

\item{error_beep}{Which beep? Defaults to "Wilhelm" (#9)}
}
\description{
beepr::beep() only on error
}
\examples{

beep_on_error(log("foo"))   # Beeps on error
beep_on_error(log(3))      # No error, no beep
}
\keyword{beep}
