% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap_a_word.R
\name{cap_a_word}
\alias{cap_a_word}
\title{Cap a Word}
\usage{
cap_a_word(phrase, to_cap = c("id", "Id"), collapse = " ")
}
\arguments{
\item{phrase}{A single element or phrase containing one or more of the elements in `to_cap`}

\item{to_cap}{A vector of words that, if contained in `phrase`, will be fully capitalized.}
}
\description{
Find a word within a word and capitalize it.
}
\examples{

cap_a_word("this_id")
cap_a_word("this_id", collapse = "_")

words_to_cap <- c("Petal", "Width")
names(iris) \%>\% map(cap_a_word, words_to_cap) \%>\% as_vector()
}
\keyword{cap}
