% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap_a_word_partial.R
\name{cap_a_word_partial}
\alias{cap_a_word_partial}
\title{Cap a Word, Partial}
\usage{
cap_a_word_partial(phrase, to_cap = c("id", "Id"))
}
\arguments{
\item{phrase}{A single element or phrase containing one or more of the elements in `to_cap`}

\item{to_cap}{A vector of words that, if contained in `phrase`, will be fully capitalized.}
}
\description{
Find part of a word and make it full caps.
}
\examples{

cap_a_word_partial("this_id")
cap_a_word_partial("this Id")
cap_a_word_partial("thisid")

words_to_cap <- c("Petal", "Width")
purrr::map(names(iris), cap_a_word_partial, words_to_cap)
}
\keyword{cap}
