% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap_df.R
\name{cap_df}
\alias{cap_df}
\title{Cap It}
\usage{
cap_df(df)
}
\arguments{
\item{df}{A dataframe.}
}
\description{
Capitalize the column names in a dataframe.
}
\examples{
mtcars_2 <- mtcars \%>\% cap_df()
}
\keyword{capitalize}
