% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize_df.R
\name{capitalize_df}
\alias{capitalize_df}
\title{Capitalize Dataframe Columns}
\usage{
capitalize_df(df, ...)
}
\arguments{
\item{df}{Any dataframe.}
}
\description{
Capitalize the column names in a dataframe.
Somewhat deprecated in favor of cap_it()
}
\examples{
capitalize_df(iris)
}
\keyword{capitalize}
