% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize_this.R
\name{capitalize_this}
\alias{capitalize_this}
\title{Beer Caps}
\usage{
capitalize_this(df, ...)
}
\arguments{
\item{df}{Any dataframe.}
}
\description{
Capitalize beer words.
}
\examples{
capitalize_this()
}
\keyword{capitalize}
