% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_strings.R
\name{collapse_strings}
\alias{collapse_strings}
\title{Collapse Strings}
\usage{
collapse_strings(vec)
}
\arguments{
\item{vec}{A vector of strings.}
}
\description{
Collapse a vector of strings into one single string separated by commas. The last
element will be preceed with an "and ,".
}
\examples{

letters \%>\% collapse_strings()
}
\keyword{collapse}
