% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_off.R
\name{cut_off_vecwise}
\alias{cut_off_vecwise}
\title{Cut Off}
\usage{
cut_off_vecwise(vec, cutoff)
}
\arguments{
\item{vec}{A numeric vector.}

\item{cutoff}{A numeric cutoff value.}

\item{e}{An element of a vector.}
}
\description{
Throw out elements that exceed a certain absolute value.
}
\examples{

mtcars$mpg \%>\% cut_off_vecwise(20)
mtcars$carb \%>\% map(cut_off_elementwise)
}
\keyword{cluster}
