% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disconnect_all.R
\name{disconnect_all}
\alias{disconnect_all}
\title{Disconnect All}
\usage{
disconnect_all(drv = RMySQL::MySQL())
}
\arguments{
\item{drv}{A db driver. Defaults to \code{RMySQL::MySQL()}.}
}
\value{
List of all closed connections.
}
\description{
Disconnect all open database connections.
}
\examples{
\dontrun{
 disconnect_all()
}
}
\keyword{database}
