% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_cluster.R
\name{do_cluster}
\alias{do_cluster}
\title{K-Means Cluster}
\usage{
do_cluster(df, vars, to_scale, n_centers = 5, algorithm = "Hartigan-Wong",
  trace = FALSE)
}
\arguments{
\item{df}{A dataframe}

\item{vars}{Variables from that dataframe to keep, some of which will be scaled and used for clustering.}

\item{to_scale}{A subset of vars, all numeric, to be scaled and used for clustering.}

\item{n_centers}{The number of clusters you want.}

\item{algorithm}{kmeans algo to use}

\item{trace}{Trace progress?}
}
\description{
From beginning to end
}
\examples{

v <- names(iris)
t_s <- v[-which(v == "Species" | v == "Petal.Length")]

clustered_iris <- do_cluster(iris, vars = v, to_scale = t_s, n_centers = 3)

ggplot2::ggplot(data = clustered_iris \%>\% trim_outliers(exclude = c("cluster_assignment", "Species"))) +
  ggplot2::geom_text(ggplot2::aes(Sepal.Width_scaled, Sepal.Length_scaled,
                colour = cluster_assignment,
                label = Species)) +
  ggplot2::theme_bw()
}
\keyword{cluster}
