% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_last_non_na_value.R
\name{find_last_non_na_value}
\alias{find_last_non_na_value}
\title{Find last non-NA value}
\usage{
find_last_non_na_value(vec)
}
\arguments{
\item{vec}{A datafame}
}
\description{
Take vector and start at the end, to find the last non-NA element
to look for by |.
}
\examples{

c(1, 2, NA_integer_, 4, 5, NA_integer_, NA_integer_) \%>\% find_last_non_na_value()
sample(c(letters, rep(NA_character_, 3))) \%>\% find_last_non_na_value()
}
\keyword{NA}
