% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nums.R
\name{find_nums}
\alias{find_nums}
\title{Contains a Number?}
\usage{
find_nums(df, col = word, add_contains_num = TRUE)
}
\arguments{
\item{df}{A dataframe}

\item{col}{The unquoted name of the column of interest}

\item{add_contains_num}{Do you just want}
}
\value{
Adds a new column for whether or not a value in col is or contains a number.
}
\description{
If something can be coerced to numeric, is_num is TRUE; if it just contains digits, contains_num is TRUE.
}
\examples{

mt_df <- mtcars \%>\% dplyr::mutate(car_name = rownames(.))
mt_df \%>\% find_nums(car_name)
}
