% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fizzbuzz.R
\name{fizzbuzz}
\alias{fizzbuzz}
\title{FizzBuzz}
\usage{
fizzbuzz(len)
}
\arguments{
\item{len}{Length of vector}

\item{i}{Index, in fiz_vec}
}
\description{
Compare for loop time to vectorized time.
}
\details{
FizzBuzz problem: Print 1 to 100; for multiples of three print "Fizz", for the multiples of five print "Buzz".
For numbers which are multiples of both three and five print "FizzBuzz".
}
\examples{


system.time(fizzbuzz(1000))
system.time(fizzbuzz_2(1000))

system.time(lapply(1:1000, fiz_vec))
system.time(purrr::map(1:1000, fiz_vec))

system.time(sapply(1:1000, fiz_vec))
system.time(purrr::map_chr(1:1000, fiz_vec))
}
\keyword{fizzbuzz}
