% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nn_importance.R
\name{get_nn_importance}
\alias{get_nn_importance}
\title{Get Neural Net Variable Importance}
\usage{
get_nn_importance(nn, keep_gini = FALSE)
}
\arguments{
\item{nn}{A neural net or multinomial neural net}

\item{keep_gini}{Keep the coefficient or just the percent importance?}
}
\description{
Given a neural net, extract the importance coefficient and percent.
}
\examples{

iris_nn <- nnet::nnet(Species ~ ., data = iris, size = 3)
iris_multinom <- nnet::multinom(Species ~ ., data = iris)
iris_nn \%>\% get_nn_importance()
iris_multinom \%>\% get_nn_importance()
}
\keyword{importance}
