% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rf_importance.R
\name{get_rf_importance}
\alias{get_rf_importance}
\title{Get Random Forest Variable Importance}
\usage{
get_rf_importance(model, keep_gini = FALSE)
}
\arguments{
\item{model}{A random forest model}

\item{keep_gini}{Keep the coefficient or just the percent importance?}
}
\description{
Given a random forest model, extract the importance coefficient and percent.
}
\examples{

library(ranger)
library(tidyverse)
iris_rf <- ranger::ranger(Species ~ ., data = iris, importance = "impurity")
iris_rf \%>\% get_rf_importance()
}
\keyword{importance}
