% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_first_word.R
\name{grab_first_word}
\alias{grab_first_word}
\title{Grab only the first word from a string.}
\usage{
grab_first_word(e, splitter = " ")
}
\arguments{
\item{splitter}{The character vector that delineates where the first word ends.}

\item{d}{A character element to split.}
}
\description{
Specify the end of the word using the splitter variable.
}
\examples{
grab_first_word("some string but I just want the first word")

dplyr::starwars \%>\%
  dplyr::select(name) \%>\%
  dplyr::mutate(
    first_name = purrr::map_chr(name, grab_first_word),
    droid_first_name = purrr::map_chr(first_name, grab_first_word, splitter = "-")
  )
}
\keyword{capitalize}
