% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_equal.R
\name{group_equal}
\alias{group_equal}
\title{Group Equal}
\usage{
group_equal(df, col, n_groups = 5, add_first_last = TRUE)
}
\arguments{
\item{df}{A dataframe}

\item{col}{A numeric column. If the value here is outside the vector range, the `group` colum
will be set to NA for that row.}

\item{n_groups}{Number of groups to end.}
}
\description{
Split a dataframe column into equal-sized groups.
}
\examples{

groups <- group_equal(mtcars, "mpg", n_groups = 5)
group_numeric(mtcars, "mpg", groups, add_first_last = TRUE)
}
\keyword{group}
