% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_numeric.R
\name{group_numeric}
\alias{group_numeric}
\title{Group Numeric}
\usage{
group_numeric(df, col, vec, add_first_last = FALSE)
}
\arguments{
\item{df}{A dataframe}

\item{col}{A numeric or date column. If the value here is outside the vector range, the `group` colum
will be set to NA for that row.}

\item{vec}{A vector of cuts.}

\item{add_first_last}{Should the first and last element of the vector be used as the lower and upper bounds,
respectively, for the 1st and nth element? Or should those
In other words, should the 1st element of the vector be considered the lower (add_first_last = FALSE)
 or upper (add_first_last = TRUE) bound of the first group?}
}
\description{
Mutate a dataframe by assigning each row a group based on its value in a certain column.
}
\examples{

mpg_cuts <- c(15, 20, 25, 30)
group_numeric(mtcars, "mpg", mpg_cuts, add_first_last = TRUE)

mtcars$some_dates <- (rnorm(n = 32) * 5) \%>\% lubridate::as_date()
date_cuts <- c("1969-11-01", "1969-12-29", "1970-01-01", "1970-01-04", "1970-02-01")
group_numeric(mtcars, "some_dates", date_cuts, add_first_last = FALSE)
}
\keyword{group}
