% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_scripts.R
\name{import_scripts}
\alias{import_scripts}
\title{Import Scripts}
\usage{
import_scripts(path, pattern = "*.R")
}
\arguments{
\item{path}{The directory path}

\item{pattern}{Extension pattern}
}
\description{
Source in a directory of R scripts
}
\examples{

droids   # droids defined in data/get_droids.R
}
