% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_plural.R
\name{is_plural}
\alias{is_plural}
\title{Is this word plural?}
\usage{
is_plural(word, return_bool = FALSE)
}
\arguments{
\item{word}{A string}

\item{return_bool}{Should you get the word to say back or a boolean?}
}
\description{
Return either the correct word to say for this string ("them" or "it") or a boolean value
}
\examples{
is_plural("CATS")
is_plural("dog", return_bool = TRUE)
}
\keyword{plural}
