% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hist.R
\name{make_hist}
\alias{make_hist}
\title{Make a Histogram of a Variable}
\usage{
make_hist(d, var, ...)
}
\arguments{
\item{d}{A dataframe.}

\item{var}{The quoted variable name you want to see a histogram of.}
}
\description{
Choose a variable from a dataframe and make a histogram out of it. Capitalizes variable names.
}
\examples{
make_hist(mtcars, "carb")
}
\keyword{plot}
