% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_maker.R
\name{match_maker_full}
\alias{match_maker_full}
\title{Fuzzy Text Matching: vector}
\usage{
match_maker_full(vec, matches, max_dist = 5)
}
\arguments{
\item{vec}{A vector}

\item{matches}{A vector in which to look for matches with e.}

\item{max_dist}{Set maxDist to be used in stringdist::amatch}
}
\description{
Find the best match (or no match at all) to string inputs.
}
\examples{

iris <- iris \%>\% tibble::as_tibble()
match_maker_full(as.character(iris$Species),
                 matches = c("Virginia", "California", "Sarasota"), max_dist = 20)
}
\keyword{match}
