% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_to_zero_df.R
\name{na_to_zero_df}
\alias{na_to_zero_df}
\title{NA to Zeros}
\usage{
na_to_zero_df(df, cols, ...)
}
\arguments{
\item{df}{A dataframe.}

\item{cols}{A vector of quoted column names.}

\item{...}{Other arguments}
}
\description{
Specify a dataframe and a vector of column names. If there are NAs in those columns,
convert them to 0s and return a dataframe. This is slower than using na_to_zero.R with purrr:: map().
}
\examples{
sample_df <- list(a = 1:3,
                  b = letters[1:3],
                  c = c(NA, "foo", "bar")) \%>\%
  tibble::as_tibble()
cols_to_zero <- c("b", "c")
na_to_zero_df(sample_df, cols_to_zero)
}
\keyword{zero}
