% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nerdR.R
\name{nerdR}
\alias{nerdR}
\title{But you already knew that}
\usage{
nerdR()
}
\description{
Say it loud and clear. At the request of one [Kyle Gilbert](https://github.com/klye-g).
}
\examples{
nerdR()
}
\keyword{nerd}
