% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_all_na.R
\name{not_all_na}
\alias{not_all_na}
\title{Not all NA}
\usage{
not_all_na(x)
}
\arguments{
\item{x}{Vector}
}
\description{
Logical for whether a vector contains all NAs
}
\examples{

c(1:3) \%>\% not_all_na()
rep(NA, 3) \%>\% not_all_na()
c(1, 2, NA) \%>\% not_all_na()
}
