% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_clusters.R
\name{prep_clusters}
\alias{prep_clusters}
\title{Prep Clusters}
\usage{
prep_clusters(df, preds, to_scale, resp)
}
\arguments{
\item{df}{A dataframe containing data to cluster on.}

\item{preds}{A vector of predictor variables.}

\item{to_scale}{A subset of those predictors that should be scaled.}

\item{resp}{The response variable.}
}
\description{
Prepare a dataframe for clustering. Specify a vector of predictors, scale some or all of them,
and specify a response variable.
Return a list with the original dataframe (df_for_clustering), the predictors (preds), and the
outcome variables (outcome).
}
\examples{



cluster_on <- c("mpg", "cyl", "disp", "hp", "drat")
to_scale <- c("mpg", "cyl", "disp")
response_vars <- c("gear", "carb")

prep_car_clusters <- prep_clusters(df = mtcars,
                              preds = cluster_on,
                              to_scale = to_scale,
                              resp = response_vars)
}
\keyword{cluster}
