% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_rf.R
\name{prep_rf}
\alias{prep_rf}
\title{Prepare a Random Forest model}
\usage{
prep_rf(data, predictor_vars, outcome_var, n_trees = 500, filter_var = NULL,
  filter_to = NULL, prob = FALSE)
}
\arguments{
\item{predictor_vars}{A vector of predictors}

\item{outcome_var}{A single outcome variable. Factor or numeric.}

\item{filter_var}{A variable to filter}

\item{filter_to}{A value in that filter variable to filter to.}

\item{prob}{Probability or response outcome? If true, outcome variable must be a factor.}
}
\description{
Given a random forest model, extract the importance coefficient and percent.
}
\examples{

preds <- c("disp", "hp", "qsec")
response <- "carb"
mt_rf <- prep_rf(mtcars, preds, response, prob = FALSE)
}
\keyword{rf}
