% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_empty_strings.R
\name{remove_empty_strings}
\alias{remove_empty_strings}
\title{Remove empty string elements}
\usage{
remove_empty_strings(vec)
}
\arguments{
\item{vec}{A vector that may contain empty strings}
}
\description{
Take a vector and remove empty string elements. Useful after string splitting and being left with lots of empty string vectors.
}
\examples{

remove_empty_strings(c("a", "", "c"))
}
\keyword{string}
