% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_html_tags.R
\name{remove_html_tags}
\alias{remove_html_tags}
\title{Remove HTML tags}
\usage{
remove_html_tags(h_str)
}
\arguments{
\item{h_str}{A string possibly containing HTML tags}
}
\value{
Cleaned string
}
\description{
Take a string and remove HTML tags from hit.
}
\examples{

to_clean <- "<li> some \\n <b>bullet</b> </li>"
remove_html_tags(to_clean)
}
