% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_names.R
\name{replace_colnames}
\alias{replace_colnames}
\title{Replace Names}
\usage{
replace_colnames(orig, new)
}
\arguments{
\item{orig}{A vector of names to replace}

\item{new}{The column name of the ref_df vector serving as the new names}
}
\description{
Replace a vector of column names with another vector
}
\examples{

replacement_names <- c("fee", "fi", "fo", "fum", "foo")
names(iris) \%>\% replace_colnames(replacement_names)
}
\keyword{names}
