% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na_df.R
\name{replace_na_df}
\alias{replace_na_df}
\title{Replace NA for a dataframe}
\usage{
replace_na_df(df, replacement = "")
}
\arguments{
\item{df}{Dataframe}

\item{replacement}{Defaults to NA_character_}
}
\description{
Replace all NA values in a dataframe
}
\examples{

starwars \%>\% replace_na_df()

}
