% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_names.R
\name{replace_names}
\alias{replace_names}
\title{Replace Names}
\usage{
replace_names(vec, to_replace, new)
}
\arguments{
\item{vec}{Vector, some or all of which you want to replace}

\item{to_replace}{Subset of vec that you want to replace}

\item{new}{Replacements for to_replace}
}
\description{
Replace a vector of column names with another vector
}
\examples{

names_to_replace <- names(iris)[1:3]
replacement_names <- c("fee", "fi", "fo")
replace_names(c("foo", names(iris), "bar"), names_to_replace, replacement_names)
}
\keyword{names}
