% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_x.R
\name{replace_x}
\alias{replace_x}
\title{Replace x}
\usage{
replace_x(x, replacement = NA_character_)
}
\arguments{
\item{x}{Element to replace}

\item{replacement}{Defaults to NA_character_}
}
\description{
Replace missing or NA values with a replacement
}
\examples{

vec <- c("a", NA, "b")
vec \%>\% map(replace_x, replacement = "foo")
}
