% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rf.R
\name{run_rf}
\alias{run_rf}
\title{Run a Random Forest model}
\usage{
run_rf(data, predictor_vars, outcome_var, n_trees = 500, prob = FALSE)
}
\arguments{
\item{predictor_vars}{A vector of predictors}

\item{outcome_var}{A single outcome variable. Factor or numeric.}

\item{prob}{Probability or response outcome? If TRUE, outcome variable must be a factor.}
}
\description{
Prepare and run a random forest model.
}
\examples{

preds <- c("disp", "hp", "qsec")
response <- "carb"
mt_rf <- run_rf(mtcars, preds, response, prob = FALSE)
}
\keyword{rf}
