% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tests.R
\name{run_tests}
\alias{run_tests}
\title{Run Tests}
\usage{
run_tests(dir = NULL, files = NULL, ext = NULL,
  line_by_line = TRUE, beep = TRUE, beep_num = 1, ...)
}
\arguments{
\item{dir}{A directory in which your test files live}

\item{files}{A vector of test files, with or without extensions}

\item{ext}{Optional extension, if does not exist in file name.}

\item{line_by_line}{Do you want to step through with prompts? If non-interactive, this is not an option}

\item{beep}{Should we beep when done?}

\item{beep_num}{If so, what beep number should we use?}

\item{...}{Args to be passed to \code{\link{test_it}}}
}
\description{
Run Tests
}
\examples{

\dontrun{
my_test_files \%>\% run_tests()
}
}
