% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_col_types.R
\name{set_col_types}
\alias{set_col_types}
\title{Set Column Types}
\usage{
set_col_types(df, fac_regex = "", num_regex = "", date_regex = "",
  char_regex = "", ...)
}
\arguments{
\item{df}{A datafame}

\item{fac_regex}{A regex of column names to be turned into factors.}

\item{num_regex}{A regex of column names to be turned numeric.}

\item{date_regex}{A regex of column names to be turned into dates.}

\item{char_regex}{A regex of column names to be turned into character.}

\item{...}{Other arguments}
}
\description{
Set column types in a dataframe using regexes. Separate full or partial column names
to look for by |.
}
\examples{
my_df <- as.data.frame(list(some = 1:5, column = Sys.Date() + 1:5,
                            types = letters[5:9],
                            to = c("good", "bad", "ugly", "other", "NA"),
                            change = rnorm(10:14)))

factor_cols <- c("change")
char_cols <- c("some|to")

my_new_df <- set_col_types(df = my_df, fac_regex = factor_cols,
                           char_regex = char_cols)
str(my_df)
str(my_new_df)
}
\keyword{column}
\keyword{types}
