% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize_this.R
\name{simple_cap}
\alias{simple_cap}
\title{Simple Capitalization}
\usage{
simple_cap(x)
}
\arguments{
\item{x}{Any vector}
}
\description{
Capitalize a vector of words.
}
\examples{
simple_cap()
}
\keyword{capitalize}
