% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{some_starwars}
\alias{some_starwars}
\title{Sample of starwars data}
\format{A data frame}
\source{
dplyr
}
\usage{
some_starwars
}
\description{
Sample of starwars data
}
\examples{
\dontrun{
 some_starwars
}
}
\keyword{datasets}
