% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_lm.R
\name{style_lm}
\alias{style_lm}
\title{Style a Regression Model Output}
\usage{
style_lm(tidied_lm)
}
\arguments{
\item{tidied_lm}{A tidied regression summary.}
}
\description{
Capitalize and round the result of a tidied regression.
Remove the Intercept term and Statistic column, capitalize all terms and names.
Round all numbers to the thousandths.
}
\examples{
iris_lm <- lm(Sepal.Width ~ Sepal.Length, data = iris) \%>\% summary() \%>\% broom::tidy()
style_lm(iris_lm)
}
\keyword{regression}
