% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_mod.R
\name{style_mod}
\alias{style_mod}
\title{Style glm or lm}
\usage{
style_mod(m)
}
\arguments{
\item{d}{A dataframe.}

\item{var}{The quoted variable name you want to see a histogram of.}
}
\description{
Style a model output depending on whether it's a linear model or generalized linear model.
}
\examples{
mt_lm <- lm(mpg ~ hp, data = mtcars)
mt_lm \%>\% style_mod()

iris_glm <- glm(Species ~ Sepal.Width, data = iris, family = binomial)
iris_glm \%>\% style_mod()
}
\keyword{capitalize}
