% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_points.R
\name{summarise_points}
\alias{summarise_points}
\title{Trello Point Summary}
\usage{
summarise_points(df, lowercase = TRUE, ...)
}
\arguments{
\item{df}{A datafame}

\item{lowercase}{"billed" and "points" by default. If FALSE, will use `Billed?` and Points}

\item{...}{Other arguments}
}
\description{
Summarise billed and unbilled points in a dataframe.
}
\examples{

library(tidyverse)
lowercase_df <- as.data.frame(list(points = seq(1, 10, by = 2),
                            billed = c("Yes", "No", "No", "Yes", "Yes")))

uppercase_df <- lowercase_df \%>\%
  rename(
    Points = points,
    `Billed?` = billed
  )

summarise_points(lowercase_df)
summarise_points(uppercase_df, lowercase = FALSE)
}
\keyword{summarise}
