% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_it.R
\name{test_it}
\alias{test_it}
\title{Test It}
\usage{
test_it(f, verbose = FALSE)
}
\arguments{
\item{f}{The file}

\item{verbose}{Should messages and warnings be suppressed}
}
\description{
Return whether an individual test passed or not
}
\examples{

}
