% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_content.R
\name{trim_content}
\alias{trim_content}
\title{Trim content}
\usage{
trim_content(df, col = content, last_char = 50, sample_some = 3)
}
\arguments{
\item{df}{Input dataframe}

\item{col}{Unquoted column name of character column}

\item{last_char}{Starting at character 1, how what's the last character that should be included.}

\item{sample_some}{Number of rows to sample. If NULL, all rows are retained.}
}
\description{
For easy kable-izing, take a sample of a dataframe and trim a single column (usually a long string)
}
\examples{

tbl <- janeaustenr::northangerabbey[100:109] \%>\%
  tibble::enframe() \%>\%
  dplyr::rename(
    content = value
  ) \%>\%
  dplyr::mutate(
    num = runif(10)
  )

tbl \%>\% trim_content(last_char = 5, sample_some = 7)
}
\keyword{trim}
