% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncap_it.R
\name{uncap_it}
\alias{uncap_it}
\title{Uncap It}
\usage{
uncap_it(e, ...)
}
\arguments{
\item{e}{A vector or dataframe.}

\item{...}{Other arguments}
}
\description{
Uncapitalize a vector or dataframe (usually the column names) and replace spaces
with underscores using one of the apply or purrr functions.
}
\examples{

string_to_tidy <- "A String that needs Tidying"
uncap_it(string_to_tidy)
}
\keyword{uncapitalize}
