% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_score.R
\name{z_score}
\alias{z_score}
\title{Z-Score a vector}
\usage{
z_score(vec)
}
\arguments{
\item{vec}{A numeric vector}
}
\description{
Scale a vector and get a plain vanilla numeric vector back.
}
\examples{

z_score(mtcars$mpg)
}
\keyword{scale}
