% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_cluster.R
\name{do_cluster}
\alias{do_cluster}
\title{K-Means Cluster}
\usage{
do_cluster(df, vars, to_scale, n_centers = 5)
}
\arguments{
\item{df}{A dataframe}

\item{vars}{Variables from that dataframe to keep, some of which will be scaled and used for clustering.}

\item{to_scale}{A subset of vars, all numeric, to be scaled and used for clustering.}

\item{n_centers}{The number of clusters you want.}
}
\description{
From beginning to end
}
\examples{

v <- names(iris)
t_s <- v[-which(v == "Species" | v == "Petal.Length")]

clustered_iris <- do_cluster(iris, vars = v, to_scale = t_s, n_centers = 3)

ggplot(data=clustered_iris) +
  geom_text(aes(Sepal.Width_scaled, Sepal.Length_scaled,
                colour = cluster_assignment,
                label = Species)) +
  theme_bw()
}
\keyword{cluster}
