% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_scripts.R
\name{import_scripts}
\alias{import_scripts}
\title{Import Scripts}
\usage{
import_scripts(path, pattern = "*.R")
}
\arguments{
\item{path}{The directory path}

\item{pattern}{Extension pattern}
}
\description{
Source in a directory of R scripts
}
\examples{

import_scripts(path = "./data")
droids   # droids defined in data/get_droids.R
import_scripts(path = "foo")   # Should fail with a warning, not an error
}
