% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_scripts.R
\name{import_scripts}
\alias{import_scripts}
\title{Import Scripts}
\usage{
import_scripts(path, pattern = "*.R", recursive = FALSE, verbose = FALSE)
}
\arguments{
\item{path}{The directory path}

\item{pattern}{Extension pattern}

\item{recursive}{Grab files in subdirectories?}

\item{verbose}{Should script be sourced quietly?}
}
\description{
Source in a directory of R scripts
}
\examples{

\dontrun{
import_scripts(path = "./data", pattern = ".rda")

# Import scripts
import_scripts(path = "./data-raw", pattern = "get_droids.R", verbose = TRUE)
import_scripts(path = "foo")   # Should fail with a warning, not an error
}
}
