% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_maker.R
\name{match_maker}
\alias{match_maker}
\title{Fuzzy Text Matching}
\usage{
match_maker(e, matches, max_dist = 5)
}
\arguments{
\item{e}{An character}

\item{matches}{A vector in which to look for matches with e.}
}
\description{
Find the best match (or no match at all) to string inputs.
}
\examples{

iris \%>\% dplyr::mutate(
  foo = purrr::map_chr(as.character(Species), match_maker,
                matches = c("Virginia", "California", "Sarasota"))
  )

iris \%>\% dplyr::mutate(
  foo = purrr::map_chr(as.character(Species), match_maker,
                matches = c("Virginia", "California", "Sarasota"), max_dist = 20)
)

match_maker_full(as.character(iris$Species),
                 matches = c("Virginia", "California", "Sarasota"), max_dist = 20)

}
\keyword{match}
