% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_to_zero.R, R/na_to_zero_df.R
\name{na_to_zero}
\alias{na_to_zero}
\title{NA to Zeros}
\usage{
na_to_zero(df, cols, ...)

na_to_zero(df, cols, ...)
}
\arguments{
\item{df}{A dataframe.}

\item{cols}{A vector of quoted column names.}

\item{val}{A value.}
}
\description{
An alternative to na_to_zero_df(). Use purrr::map() instead.

Specify a dataframe and a vector of column names. If there are NAs in those columns,
convert them to 0s and return a dataframe. This is slower than using na_to_zero.R with purrr:: map().
}
\examples{
library(tidyverse)
sample_df <- list(a = c(1:2, NA), b = letters[1:3], c = c(NA, "foo", "bar")) \%>\% as_tibble()
sample_df \%>\% map_df(na_to_zero)

cols_to_zero <- c("a", "b")
sample_df \%>\% map_at(na_to_zero, .at = cols_to_zero) \%>\% as_tibble()
sample_df <- list(a = 1:3, b = letters[1:3], c = c(NA, "foo", "bar")) \%>\% as_tibble()
cols_to_zero <- c("b", "c")
na_to_zero(sample_df, cols_to_zero)
}
\keyword{zero}
