% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_to_zero.R
\name{na_to_zero}
\alias{na_to_zero}
\title{NA to Zeros}
\usage{
na_to_zero(val)
}
\arguments{
\item{val}{A value.}
}
\description{
An alternative to na_to_zero_df(). Use purrr::map() instead.
}
\examples{

sample_df <- list(a = c(1:2, NA),
                  b = letters[1:3],
                  c = c(NA, "foo", "bar")) \%>\%
  tibble::as_tibble()
sample_df \%>\% purrr::map_df(na_to_zero)

cols_to_zero <- c("a", "b")
sample_df \%>\% purrr::map_at(na_to_zero, .at = cols_to_zero) \%>\% tibble::as_tibble()
}
\keyword{zero}
