% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tests.R
\name{run_tests}
\alias{run_tests}
\title{Run Tests}
\usage{
run_tests(files = NULL, ext = NULL, line_by_line = TRUE)
}
\arguments{
\item{files}{A vector of test files, with or without extensions}

\item{ext}{Optional extension, if does not exist in file name.}

\item{line_by_line}{Do you want to step through with prompts? If non-interactive, this is not an option}
}
\description{
Run Tests
}
\examples{

my_test_files \%>\% run_tests()

}
