% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_lm.R
\name{style_lm}
\alias{style_lm}
\title{Style a Regression Model Output}
\usage{
style_lm(tidied_lm, add_commas = FALSE)
}
\arguments{
\item{tidied_lm}{A tidied regression summary.}

\item{add_commas}{Should commas be added to numbers >= 1000?}
}
\description{
Capitalize and round the result of a tidied regression.
Remove the Intercept term and Statistic column, capitalize all terms and names.
Round all numbers to the thousandths.
}
\examples{
iris_lm <- lm(Sepal.Width ~ Sepal.Length, data = iris) \%>\% summary() \%>\% broom::tidy()
iris_lm[2, 3] <- 1234567
style_lm(iris_lm, add_commas = TRUE)
}
\keyword{regression}
