% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_mod.R
\name{style_mod}
\alias{style_mod}
\title{Style glm or lm}
\usage{
style_mod(m, add_commas = FALSE)
}
\arguments{
\item{m}{The output (not summarized) of a generalized linear model (glm) or linear model (lm).}

\item{add_commas}{Should commas be added to numbers >= 1000?}
}
\description{
Style a model output depending on whether it's a linear model or generalized linear model.
}
\examples{
mt_lm <- lm(mpg ~ hp, data = mtcars)
mt_lm \%>\% style_mod()

iris_glm <- glm(Species ~ Sepal.Width, data = iris, family = binomial)
iris_glm \%>\% style_mod()
}
\keyword{style}
