% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_outliers.R
\name{trim_outliers}
\alias{trim_outliers}
\title{Trim Outliers}
\usage{
trim_outliers(df, cutoff = 1.96, exclude = NULL, keep_scaled = TRUE)
}
\arguments{
\item{df}{A dataframe.}

\item{cutoff}{A numeric standard deviation cutoff value.}

\item{exclude}{Vector of variables not to subject to cutting off.}

\item{keep_scaled}{Should the newly scaled numeric variables be kept in the output?}
}
\description{
Remove outliers
}
\examples{

trim_outliers(iris, cutoff = 1, exclude = c("Species", "Sepal.Length"), keep_scaled = FALSE)
}
\keyword{trim}
